requirements.txt
Pillow
pynput 